requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
spotipy>=2.23.0
